requirements.txt
Django==3.0.6
mysqlclient==1.4.6
celery==4.4.2
Pillow==7.1.2
opencv-python==4.2.0.34
ffmpeg-python==0.2.0